extract BlogPostSummary
